Fixing tests

+++ src/main/java/com/example/trading_system/domain/stores/MarketFacadeImp.java
@@ -41,8 +41,10 @@ public class MarketFacadeImp implements MarketFacade {
     @Override
     public void deleteInstance() {
         if (instance != null) {
-            instance.storeRepository.deleteInstance();
-            instance.storeRepository = null;
+            if(instance.storeRepository==null){
+                instance.storeRepository.deleteInstance();
+                instance.storeRepository = null;
+            }
             instance.userFacade = null;
             instance = null;
         }
